// src/app/models/store.model.ts
export interface StoreModel {
    id: string;
    name: string;
    serviceTypes: string[];
    properties: {
        callWaiter: boolean;
        callWallet: boolean;
        requestCheck: boolean;
        reservation: boolean;
        invoiceActive: boolean;
        pay: boolean;
        order: boolean;
        webMenuActive: boolean;
        webAppActive: boolean;
        timedDineIn: boolean;
        timedTakeAway: boolean;
        delivery: boolean;
        selfOrdering: boolean;
        selfService: boolean;
        stampActive: boolean;
        stampConversionRate: number;
    };
    languages: string[];
    localPayTypes: string[];
    payTypes: string[];
    cloudId: string;
    mainAction: string;
    menuVersion: number;
    orwiPayPoint: number;
    payPoint: number;
    kerzzId: string;
    tipEnabled: boolean;
    workingHours: Array<{
        day: number;
        open: string;
        close: string;
        closed: boolean;
    }>;
    isProduction: boolean;
    storeCurrency: string;
    storeLocale: string;
    parentId: string;
    pointTitle: string;
    maxPoint: number;
    isVisible: boolean;
    sharedPoints: boolean;
    deliveryStoreId: string;
    address: string;
    phone: string;
    location: string;
    geoLocation: {
        latitude: number;
        longitude: number;
    };
    instagramAddress: string;
    amountPerPerson: number;
    description: string;
    rate: number;
    mainCategoryId: string;
    status: string;
    subCategories: string[];
    attributes: string[];
    payoffGroupId: string;
    apiKey: string;
    benefitType: string;
    serviceFee: number;
    branchCode: number;
}